refactor: consolidate shared styles into a single file
This refactoring consolidates shared styles from various components into a single file called sharedStyles.ts. This improves code organization, reduces duplication, and makes it easier to maintain and update shared styles across components. The changes include moving common style definitions like button, text, input, and icon-related styles to the new shared file and updating import statements in affected components.
